fix: wallet dialog reset on close
